format

+++ contracts/kSwapPool.sol
@@ -13,24 +13,25 @@ import "./Math.sol";
 //A lock is used to prevent re-entrancy during the callback.
 //For simplicity , there are no Fees
 contract kSwapPool is IkSwapPool, IERC20 {
-    bytes4 private constant SELECTOR = bytes4(keccak256(bytes('transfer(address,uint256)')));
+    bytes4 private constant SELECTOR = bytes4(keccak256(bytes("transfer(address,uint256)")));
 
     error TransferFailed();
     error ContractLocked();
     error InsufficientLiquidity();
     error InsufficientLiquidityMinted();
     error InsufficientLiquidityBurned();
-    error InsufficientInputAmount();    
-    error SwapOverflow();        
+    error InsufficientInputAmount();
+    error SwapOverflow();
     error AmountIn0();
-    error InvalidTo();    
-    function _safeTransfer(address token, address to, uint value) private {
+    error InvalidTo();
+
+    function _safeTransfer(address token, address to, uint256 value) private {
         (bool success, bytes memory data) = token.call(abi.encodeWithSelector(SELECTOR, to, value));
-        require(success && (data.length == 0 || abi.decode(data, (bool))), 'KSWAP: TRANSFER_FAILED');
+        require(success && (data.length == 0 || abi.decode(data, (bool))), "KSWAP: TRANSFER_FAILED");
     }
 
-    event Mint(address indexed sender, uint amount0, uint amount1);
-    event Burn(address indexed sender, uint amount0, uint amount1, address indexed to);    
+    event Mint(address indexed sender, uint256 amount0, uint256 amount1);
+    event Burn(address indexed sender, uint256 amount0, uint256 amount1, address indexed to);
 
     event Swap(address user, int256 amount0, int256 amount1);
     //UniswapV2 like storage .  2 tokens, 2 reserves.
@@ -108,19 +109,13 @@ contract kSwapPool is IkSwapPool, IERC20 {
         return true;
     }
 
-    function getAmountOut(uint256 amountIn, uint256 reserveIn, uint256 reserveOut)
-        internal
-        pure
-        returns (
-            uint256
-        )
-    {
-        if (reserveIn == 0 ) {
+    function getAmountOut(uint256 amountIn, uint256 reserveIn, uint256 reserveOut) internal pure returns (uint256) {
+        if (reserveIn == 0) {
             revert InsufficientLiquidity();
         }
-        if (reserveOut == 0 ) {
+        if (reserveOut == 0) {
             revert InsufficientLiquidity();
-        }        
+        }
 
         uint256 numerator = amountIn * reserveOut;
         uint256 denominator = reserveIn + amountIn;
@@ -159,24 +154,23 @@ contract kSwapPool is IkSwapPool, IERC20 {
         emit Mint(msg.sender, amount0, amount1);
     }
 
-   // this low-level function should be called from a contract which performs important safety checks
-    function burn(address to) external lock returns (uint amount0, uint amount1) {
-        
-        address _token0 = token0;                                // gas savings
-        address _token1 = token1;                                // gas savings
-        uint balance0 = IERC20(_token0).balanceOf(address(this));
-        uint balance1 = IERC20(_token1).balanceOf(address(this));
-        uint liquidity = balanceOf[address(this)];
+    // this low-level function should be called from a contract which performs important safety checks
+    function burn(address to) external lock returns (uint256 amount0, uint256 amount1) {
+        address _token0 = token0; // gas savings
+        address _token1 = token1; // gas savings
+        uint256 balance0 = IERC20(_token0).balanceOf(address(this));
+        uint256 balance1 = IERC20(_token1).balanceOf(address(this));
+        uint256 liquidity = balanceOf[address(this)];
 
-        uint _totalSupply = totalSupply; // gas savings, must be defined here since totalSupply can update in _mintFee
-        amount0 = (liquidity  * balance0) / _totalSupply; // using balances ensures pro-rata distribution
+        uint256 _totalSupply = totalSupply; // gas savings, must be defined here since totalSupply can update in _mintFee
+        amount0 = (liquidity * balance0) / _totalSupply; // using balances ensures pro-rata distribution
         amount1 = (liquidity * balance1) / _totalSupply; // using balances ensures pro-rata distribution
         if (amount0 == 0) {
             revert InsufficientLiquidityBurned();
         }
         if (amount1 == 0) {
             revert InsufficientLiquidityBurned();
-        }        
+        }
 
         _burn(address(this), liquidity);
         _safeTransfer(_token0, to, amount0);
@@ -190,10 +184,8 @@ contract kSwapPool is IkSwapPool, IERC20 {
         emit Burn(msg.sender, amount0, amount1, to);
     }
 
-
     // this low-level function should be called from a contract which performs important safety checks
     function swap(address to, bool zeroForOne, uint256 amountIn, bytes calldata data) external lock returns (uint256) {
-
         if (amountIn == 0) {
             revert AmountIn0();
         }
@@ -205,7 +197,7 @@ contract kSwapPool is IkSwapPool, IERC20 {
         }
         if (to == token1) {
             revert InvalidTo();
-        }        
+        }
 
         (int256 amount0_, int256 amount1_) = zeroForOne
             ? (int256(amountIn), -int256(getAmountOut(amountIn, _reserve0, _reserve1)))
@@ -217,9 +209,9 @@ contract kSwapPool is IkSwapPool, IERC20 {
         //erc20 transfer output token
 
         if (zeroForOne) {
-            _safeTransfer(token1, to,uint256(-amount1_));
+            _safeTransfer(token1, to, uint256(-amount1_));
         } else {
-            _safeTransfer(token0, to,uint256(-amount0_));            
+            _safeTransfer(token0, to, uint256(-amount0_));
         }
 
         IkSwapSwapCallback(msg.sender).kSwapCallback(amount0_, amount1_, data);
@@ -228,11 +220,11 @@ contract kSwapPool is IkSwapPool, IERC20 {
         uint256 balanceOut = IERC20(inToken).balanceOf(address(this));
 
         if (balanceIn + amountIn > balanceOut) {
-            revert  InsufficientInputAmount();
+            revert InsufficientInputAmount();
         }
-        
+
         uint256 balance1 = IERC20(outToken).balanceOf(address(this));
-        if(balanceOut > type(uint112).max || balance1 > type(uint112).max) {
+        if (balanceOut > type(uint112).max || balance1 > type(uint112).max) {
             revert SwapOverflow();
         }
